Backend/requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.23.1
pydantic==2.5.2          # prebuilt wheel, no Rust compilation needed
python-dotenv==1.0.1
requests==2.31.0          # avoid yanked 2.32.0
PyGithub==1.59.1
motor==3.7.1
passlib[bcrypt]==1.7.4
python-jose==3.3.0
